Last version rescued.
First committed on Mar 26, 2009 (e47f0d93871c561b0f8be466609f5d80b)

  3488  baxter
  2840  megan
  2376  mike
  1953  timo
  1247  thomas
  1239  remi
  1082  david
   698  cthompson
   469  gregory
   442  brenna
   349  pdrapeau
   247  nathan
   236  sean
   193  daria
   183  bruno
   173  curtis
   135  tbone
   123  travis
   109  aaron

git-svn-id: file:///var/svn/nexopia/trunk@23069 7be605d3-0650-4ba7-bb8c-0af3a60b5258

+++ php-site/logout.php
@@ -1,6 +1,7 @@
 <?
 
 	$login = 0.5;
+	$forceserver = true;
 
 	require_once("include/general.lib.php");
 
@@ -8,12 +9,23 @@
 	$referer = getREQval('referer','string',"http://$wwwdomain/");
 
 	if($_SERVER["REQUEST_METHOD"] == "POST" || checkKey($userData['userid'], $k)){
+		// This is needed for a rare case where you let your session timeout, click the logout link, and then click the
+		//  "Click here to logout link". If a user follows those steps they will try to log out an anonymous user.
+		if(isset($userData['userid']) && isset($userData['sessionkey']) && $userData['userid'] > 0)
+		{
+			$auth->destroySession($userData['userid'], $userData['sessionkey']);
 
-		$auth->destroySession($userData['userid'], $userData['sessionkey']);
-
-		$auth->logout($userData['userid']);
-
-		header("location: $referer");
+			$auth->logout($userData['userid']);
+			
+			header("location: $referer");
+		}
+		else
+		{
+			incHeader();
+			echo "<p class=body>Congratulations! You've successfully logged out.</p>";
+			incFooter();
+		}
+	
 	}else{
 
 		incHeader();
